removed unnecessary assignment to type double

+++ Tasks_5/Task_5_2/Task_5_2.c
@@ -38,7 +38,7 @@ double performOperation(double firstNumber, double secondNumber, char operator)
 
         case '/':
         {
-            return (double)firstNumber / (double)secondNumber;
+            return firstNumber / secondNumber;
             break;
         }
